@@ -28,18 +28,16 @@ public struct ApplicationInfoView: View {
 }
 
 @available(macOS 11.0, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
-struct ApplicationInfoView_Previews: PreviewProvider {
-    static var previews: some View {
-        ApplicationInfoView()
-            .environment(\.appInfo,
-                          .init(identifier: "de.sersoft.testapp",
-                                names: .init(unlocalized: (base: "TestApp",
-                                                           display: "Test App"),
-                                             localized: (nil, nil)),
-                                versioning: .init(version: "1.0.0", build: "1")
-                               ))
-            .environment(\.appIconMode,
-                          .prerendered(Image(systemName: "squareshape.fill")))
-    }
+#Preview {
+    ApplicationInfoView()
+        .environment(\.appInfo,
+                      .init(identifier: "de.sersoft.testapp",
+                            names: .init(unlocalized: (base: "TestApp",
+                                                       display: "Test App"),
+                                         localized: (nil, nil)),
+                            versioning: .init(version: "1.0.0", build: "1")
+                           ))
+        .environment(\.appIconMode,
+                      .prerendered(Image(systemName: "squareshape.fill")))
 }
 #endif
